pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "silvergrain"
version = "0.1.0"
description = "Physically-based film grain renderer image processing"
readme = "README.md"
requires-python = ">=3.1"
license = {text = "MIT"}
authors = [
    {name = "SilverGrain Contributors"}
]
keywords = ["film-grain", "image-processing", "data-augmentation", "machine-learning", "gpu"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "numba==0.62.1",
    "numpy==2.2.6",
    "opencv-python",
    "Pillow",
]

[project.optional-dependencies]
gpu = [
    "numba-cuda[cu12]==0.20.1",
]
dev = [
    "pytest>=6.0",
    "pytest-cov",
]

[project.scripts]
silvergrain = "silvergrain.cli:main"
silvergrain-batch = "silvergrain.batch:main"

[project.urls]
Homepage = "https://github.com/yourusername/silvergrain"
Issues = "https://github.com/yourusername/silvergrain/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["silvergrain*"]
